requirements.txt
ccxt==4.0.104
pandas==2.1.1
ta==0.10.2
python-telegram-bot==20.3
requests==2.31.0
python-dotenv==1.0.0